# refactor: py store implementation with GIL management and RAII patterns