… mooncake-store/src/client.cpp
… mooncake-store/tests/test_distributed_object_store.py