performance results

main

```
Concurrent Stress Test Results:
Total threads: 8
Operations per thread: 100
Total operations: 800
Data block size: 1.00MB
Total data processed: 0.78GB
Put duration: 2.94 seconds
Get duration: 2.53 seconds
System Put throughput: 271.87 ops/sec
System Get throughput: 315.80 ops/sec
System Put bandwidth: 0.27 GB/sec
System Get bandwidth: 0.31 GB/sec
```

pr

```
Concurrent Stress Test Results:
Total threads: 8
Operations per thread: 100
Total operations: 800
Data block size: 1.00MB
Total data processed: 0.78GB
Put duration: 0.70 seconds
Get duration: 0.43 seconds
System Put throughput: 1149.84 ops/sec
System Get throughput: 1845.99 ops/sec
System Put bandwidth: 1.12 GB/sec
System Get bandwidth: 1.80 GB/sec
```